api_gateway/requirements.txt
flask
flask_cors
kafka-python